// contracts/generated/typescript/models/BulkReviewBatch.ts
export interface BulkReviewBatch {
    /**
     *
     * @type {string}
     * @memberof BulkReviewBatch
     */
    id: string;
    /**
     *
     * @type {string}
     * @memberof BulkReviewBatch
     */
    wikiId: string;
    /**
     *
     * @type {string}
     * @memberof BulkReviewBatch
     */
    createdBy: string;
    /**
     *
     * @type {BulkReviewBatchStatusEnum}
     * @memberof BulkReviewBatch
     */
    status: BulkReviewBatchStatusEnum;
    /**
     *
     * @type {BulkReviewBatchSamplingModeEnum}
     * @memberof BulkReviewBatch
     */
    samplingMode: BulkReviewBatchSamplingModeEnum;
    /**
     *
     * @type {number}
     * @memberof BulkReviewBatch
     */
    samplePercent: number;
    /**
     *
     * @type {BulkReviewBatchForceFullReasonEnum}
     * @memberof BulkReviewBatch
     */
    forceFullReason?: BulkReviewBatchForceFullReasonEnum;
    /**
     *
     * @type {number}
     * @memberof BulkReviewBatch
     */
    waveSize: number;
    /**
     *
     * @type {number}
     * @memberof BulkReviewBatch
     */
    currentWave: number;
    /**
     *
     * @type {Date}
     * @memberof BulkReviewBatch
     */
    createdAt: Date;
    /**
     *
     * @type {Date}
     * @memberof BulkReviewBatch
     */
    finalizedAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof BulkReviewBatch
     */
    pausedAt?: Date;
    /**
     *
     * @type {Date}
     * @memberof BulkReviewBatch
     */
    completedAt?: Date;
    /**
     *
     * @type {Array<BulkReviewItem>}
     * @memberof BulkReviewBatch
     */
    items: Array<BulkReviewItem>;
}


/**
 * @export
 */
export const BulkReviewBatchStatusEnum = {
    Reviewing: 'reviewing',
    Ready: 'ready',
    Applying: 'applying',
    Paused: 'paused',
    Completed: 'completed'
} as const;
export type BulkReviewBatchStatusEnum = typeof BulkReviewBatchStatusEnum[keyof typeof BulkReviewBatchStatusEnum];

/**
 * @export
 */
export const BulkReviewBatchSamplingModeEnum = {
    Sampled: 'sampled',
    Full: 'full'
} as const;
export type BulkReviewBatchSamplingModeEnum = typeof BulkReviewBatchSamplingModeEnum[keyof typeof BulkReviewBatchSamplingModeEnum];

/**
 * @export
 */
export const BulkReviewBatchForceFullReasonEnum = {
    HighOrCriticalRisk: 'high_or_critical_risk',
    ForcedByReviewer: 'forced_by_reviewer'
} as const;
export type BulkReviewBatchForceFullReasonEnum = typeof BulkReviewBatchForceFullReasonEnum[keyof typeof BulkReviewBatchForceFullReasonEnum];